[REF] Move argument handling into separate header
